Implement shader library for different backends.
